Refactored and organized the server directory structure according to separation of concerns
