+++ server/server.js
@@ -1,34 +1,14 @@
-import express from "express";
 import dotenv from "dotenv";
-import cors from "cors";
 import connectDB from "./config/db.js";
-import routes from "./routes/index.js";
-import authRoutes from "./routes/authRoutes.js";
+import app from "./app.js";
 
 dotenv.config();
-
-const app = express();
 
 // Connect to MongoDB
 connectDB();
 
-// Middleware to parse the request body as JSON
-app.use(express.json());
-
-// Middleware to allow cross-origin requests
-app.use(
-  cors({
-    origin: "http://localhost:5173", // Allow requests from the client,
-    methods: ["GET", "POST", "PUT", "DELETE"],
-    credentials: true,
-  })
-);
-
-// Use the imported routes
-app.use("/", routes);
-app.use("/api/auth", authRoutes);
-
 const PORT = process.env.PORT || 5000;
+
 app.listen(PORT, () => {
   console.log(`Server running on port: ${PORT}`);
 });
